project-6/pyproject.toml
[tool.poetry]
name = "project-6"
version = "0.1.0"
description = ""
authors = ["GbotemiB <gbotemibolarinwa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.23"
pandas = "^2.1.4"
pytest-cov = "^4.1.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pyarrow = "^14.0.1"
ipykernel = "^6.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
